+++ src/commands/setup.ts
@@ -1,10 +1,12 @@
-import { ActionRowBuilder, ButtonBuilder, ButtonStyle, ChatInputCommandInteraction, resolveColor, User } from 'discord.js'
+import { ActionRowBuilder, APIEmbed, ButtonBuilder, ButtonStyle, ChatInputCommandInteraction, GuildBasedChannel, Message, resolveColor, User } from 'discord.js'
 import Command from '../struct/Command.js'
-import Guild from '../struct/Guild.js'
+import Guild, { GuildInterface } from '../struct/Guild.js'
 import Bot from '../struct/Client.js'
 import Helper, { HelperInterface } from '../struct/Helper.js'
 import Difficulty from '../struct/Difficulty.js'
 import Responses from '../utils/responses.js'
+import { APIEmbedField } from 'discord-api-types/v9'
+import { parseMessageUrl } from '../utils/parseMessageUrl.js'
 
 export default new Command({
     name: 'setup',
@@ -151,6 +153,18 @@ export default new Command({
         {
             name: 'openapplicationmessage',
             description: 'Setup the open application message the channel',
+        },
+        {
+            name: 'formattingmsg',
+            description: ' Make or update the existing formatting message in the channel',
+            args: [
+                {
+                    name: 'optionaltext',
+                    description: 'Optional text to be inserted after the default instructions text',
+                    optionType: 'string',
+                    required: false
+                }
+            ]
         },
         {
             name: 'setuphelper',
@@ -327,6 +341,10 @@ export default new Command({
             let embed = await getSettingsEmbed(i, client, i.guild.id, 'Current server setup info') 
             return i.editReply({embeds: [embed]})  
         } else if(subCommand == 'openapplicationmessage') {
+            if(!guildData.accentColor)
+                return Responses.embed(i, '**You must first set the accent color via `/setup settings accentcolor`**')
+
+
             const applyButton = new ButtonBuilder()
                 .setCustomId('openapplicationbutton')
                 .setLabel('Apply')
@@ -336,13 +354,77 @@ export default new Command({
             
             
             const embed = {
-                title: 'Builder Applications',
-                description: 'Click the button below to open a Builder Application and apply for our South European Builder System',
+                title: 'Apply to become a builder 👷',
+                description: `Click the button below to apply to *${guildData.name}*`,
                 color: resolveColor(`#${guildData.accentColor}`)
             }
 
-            await i.channel.send({embeds: [embed], components: [row.toJSON()]})
-            return i.deleteReply()
+            let msg = await i.channel.send({embeds: [embed], components: [row.toJSON()]})
+            return Responses.embed(i, `**Open application message posted:** ${msg.url}`, guildData.accentColor)
+        } else if(subCommand == 'formattingmsg') {
+            try{
+                if(!i.channel.isSendable())
+                    return Responses.embed(i, '**Cannot send messages to this channel**')
+
+                if(!guildData.accentColor)
+                    return Responses.embed(i, '**You must first set the accent color via `/setup settings accentcolor`**')
+
+                //Check if the submissions channel is not set 
+                if(!guildData.submitChannel)
+                    return Responses.embed(i, '**Submission channel must first be set via `/setup settings submitchannel`**', guildData.accentColor)
+
+                //Check if submissions channel still exists
+                let submitChannel = await i.guild.channels.fetch(guildData.submitChannel)
+                if(!submitChannel)
+                    return Responses.embed(i, `**Could not find submission channel with ID:** ${guildData.submitChannel}`, guildData.accentColor)
+
+                let optionalText = options.getString('optionaltext')
+
+                let embed = createSubmissionFormatMessageEmbed(submitChannel, guildData, optionalText)
+
+                let msg: Message 
+
+                if(guildData.formattingMsg) {
+                    //Update the existing message if it still exists
+                    let channelMessageID = parseMessageUrl(guildData.formattingMsg)
+                    //Fetch channel where the message is posted
+                    let formattingMsgChannel = await i.guild.channels.fetch(channelMessageID.channelID) 
+                    if(formattingMsgChannel && formattingMsgChannel.isTextBased()) {
+
+                        try {
+                            let orgMsg = await formattingMsgChannel.messages.fetch(channelMessageID.messageID)
+                            //If the message exist but the id of the channel where It's posted is different from where this command was executed
+                            //delete the message. This way only one formatting message should exits per guild
+                            if(formattingMsgChannel.id != i.channelId) {
+                                await orgMsg.delete()
+                            }else {
+                                msg = orgMsg
+                            }
+                        }catch(err){}
+
+                        //Update the message
+                        if(msg)
+                            await msg.edit({embeds: [embed]})
+                    }
+                }
+
+                //Else if doesn't create it
+                if(!msg) {
+                    msg = await i.channel.send({embeds: [embed]})
+                }
+
+                //Insert the message url into the guild data
+                await Guild.updateOne({id: guildData.id}, {$set: { 'formattingMsg': msg.url}})
+                
+
+                guildData.formattingMsg = msg.url
+                //Update the client guilds data with the new formatting message url
+                client.guildsData.set(guildData.id, guildData)
+
+                return Responses.embed(i, `**Submission format message posted:** ${msg.url}`)
+            } catch(err) {
+                return Responses.embed(i, `**[FormattingMsgException]:** \n${err}}`)
+            }
         } else if(subCommand == 'setuphelper' || subCommand == 'helperstatus') {
             const user : User = options.getUser('user')
             const guildUser = await i.guild.members.fetch(user) 
@@ -429,16 +511,16 @@ function addRemoveRole(addrole: string, removerole: string, roles: string[], rol
     return hasInputData;
 }
 
-async function getSettingsEmbed(i : ChatInputCommandInteraction, client: Bot, guildId: string, msg: string) : Promise<Object> {
+async function getSettingsEmbed(i : ChatInputCommandInteraction, client: Bot, guildId: string, msg: string) : Promise<APIEmbed> {
     const guildData = await Guild.findOne({id: guildId})
     client.guildsData.set(guildId, guildData)
 
-    let embed = {
+    let embed : APIEmbed = {
         title: 'Server settings',
         description: msg
     }
 
-    let fields = []
+    let fields : APIEmbedField[] = []
 
     if(guildData.name) {
         fields.push({ name: 'Server name', value: (guildData.emoji) ? `${guildData.emoji} - ${guildData.name} - ${guildData.emoji}` : `${guildData.name}` })
@@ -525,9 +607,64 @@ async function getSettingsEmbed(i : ChatInputCommandInteraction, client: Bot, gu
         }
     }
 
-    (embed as any).fields = fields
+    embed.fields = fields
 
     return embed
+}
+
+function createSubmissionFormatMessageEmbed(submitChannel: GuildBasedChannel, guildData: GuildInterface, optionalText?: string) {
+    let paragraph1 = `To submit a build in ${submitChannel}, the submission must be in a single message in the following format:`
+
+    let messageFormat = `\`\`\`[Build count]\n[Geographic coordinates]\n[Build name (opt.), International address]\n[Collaborators (Optional)]\n[Image(s) of build]\`\`\``
+
+    let paragraph2 : string[] = ['- The build count is informative only and may not always be used directly',
+    '- Geographic coordinates are optimally within the plot bounds, but in front of the build, ex. as if it were used to set a warp',
+    '- The third line may optionally contain the build name, followed by the mandatory international address',
+    '- The optional fourth line must be set if multiple builders built a build, hence only the main builder may submit this build. It may contain one or more Discord users tagged, Minecraft usernames or simply the number of collaborators, all separated by a whitespace',
+    '- The submission must include one or more attached screenshots of the build']
+
+    let paragraph3 = `**Example submission message:**
+    > 1
+    > 41.32960553669458, 19.818597583374636
+    > Build name, Street name 1, City PostalCode, Country
+    > \@discordusername minecraftusername1 minecraftusername2 1`
+
+    let paragraph4 : string[] = ['After the submission message gets sent, it gets re-posted, hence any typos or collaborator changes can be edited later, but not before they are claimed for review with the \`/submission\` command:',
+    '- \`/submission edit [submissionIndex] <address> <coords> <buildcount>\` - Edit one or more properties',
+    '- \`/submission collaborators [submissionIndex] <add> <remove>\` - Add or remove collaborators ']
+
+    let paragraph5 = [
+        'Ex. to change the build count to 2, remove the collaborator *minecraftusername1* and remove 1 more collaborator on a submission with an index of #1, you would use two commands:',
+        '\`\`\`',
+        '/submission edit 1 buildcount: 2',
+        '/submission collaborators 1 remove: minecraftusername1 1',
+        '\`\`\`'
+    ]
+
+    let paragraph6 = `> ℹ️ **Note:** If a submission was already claimed for review but not yet reviewed, and you want to change the submission, contact the submission reviewer.`
+
+    let description = `${paragraph1}\n\n${messageFormat}\n\n${paragraph2.join('\n')}\n\n${paragraph3}\n\n${paragraph4.join('\n')}\n\n${paragraph5.join('\n')}\n\n${paragraph6}`
+
+    if(optionalText)
+        description += `\n\n${optionalText}`
+
+    let embed : APIEmbed = {
+        title: `Submission Format 📄`,
+        description: description,
+        color: resolveColor(`#${guildData.accentColor}`)
+    }
+
+    return embed
+/*
+
+
+
+
+
+
+
+
+*/
 }
 
 export { getSettingsEmbed }