+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     - Added `/plot` command to manage plots
     - Added `/accept` command to accept candidates/non-candidates as builders
     - Added `/submission` command to enable builders to edit submissions before they are claimed for review ([#2](https://github.com/BTE-Balkans/SEBS-BOT/issues/2))
+    - Added `/setup formattingmsg` sub-command to post or update the submission format message ((#3)[https://github.com/BTE-Balkans/SEBS-BOT/issues/3])
     - Added helper roles
     - Plots are added via the plots channel as messages, which can be managed/assigned via the action buttons in said message
     - Applying for a builder prompts and existing builder for their Minecraft username and registers them as one, else it creates a builder application for it